accounted for directory refactor
