+++ src/app/login/page.js
@@ -1,6 +1,6 @@
 "use client";
 
-import Login from "@/components/Login";  // using absolute path (Next.js alias @/)
+import Login from "@/components/login/Login";  // using absolute path (Next.js alias @/)
 
 export default function LoginPage() {
   return <Login />;
