+++ src/app/signup/page.js
@@ -1,6 +1,6 @@
 "use client";
 
-import Signup from "@/components/Signup";
+import Signup from "@/components/signup/Signup";
 
 export default function SignupPage() {
   return <Signup />;
